fix: put coupon methods in tickets api, not coupon
