+++ src/hooks/index.ts
@@ -42,7 +42,6 @@ import useTickets from "./apiHooks/useTickets";
 import useDonatedToday from "./apiHooks/useDonatedToday";
 import useReports from "./apiHooks/useReports";
 import useRibonConfig from "./apiHooks/useRibonConfig";
-import useCoupons from "./apiHooks/useCoupons";
 
 type InitializeHooksProps = {
   initializeApiOptions: InitializeApiProps;
@@ -87,7 +86,6 @@ export {
   useChains,
   useImpressionCards,
   useContributions,
-  useCoupons,
   useSubscriptions,
   useUserConfig,
   useUserV1Donations,
